+++ catroidTest/src/org/catrobat/catroid/uitest/formulaeditor/FormulaEditorUserVariableFragmentTest.java
@@ -167,7 +167,7 @@ public class FormulaEditorUserVariableFragmentTest extends BaseActivityInstrumen
 		UserVariable userVariable = (UserVariable) listView.getItemAtPosition(0);
 		Double setVariableToValue = Double.valueOf(SET_USERVARIABLE_TO_BRICK_VALUE);
 		assertTrue("Value of UserVariable not saved after stage1!",
-				userVariable.getValue().compareTo(setVariableToValue) == 0);
+				((Double) userVariable.getValue()).compareTo(setVariableToValue) == 0);
 
 		solo.goBack();
 		solo.waitForFragmentByTag(FormulaEditorFragment.FORMULA_EDITOR_FRAGMENT_TAG);
@@ -194,7 +194,7 @@ public class FormulaEditorUserVariableFragmentTest extends BaseActivityInstrumen
 		userVariable = (UserVariable) listView.getItemAtPosition(0);
 		setVariableToValue = Double.valueOf(SET_USERVARIABLE_TO_BRICK_VALUE);
 		assertTrue("Value of UserVariable not saved after stage2!",
-				userVariable.getValue().compareTo(setVariableToValue) == 0);
+				((Double) userVariable.getValue()).compareTo(setVariableToValue) == 0);
 
 		solo.goBack();
 		solo.waitForFragmentByTag(FormulaEditorFragment.FORMULA_EDITOR_FRAGMENT_TAG);
@@ -227,7 +227,7 @@ public class FormulaEditorUserVariableFragmentTest extends BaseActivityInstrumen
 		userVariable = (UserVariable) listView.getItemAtPosition(0);
 		setVariableToValue = Double.valueOf(SET_USERVARIABLE_TO_BRICK_VALUE);
 		assertTrue("Value of UserVariable not saved after stage3!",
-				userVariable.getValue().compareTo(setVariableToValue) == 0);
+				((Double) userVariable.getValue()).compareTo(setVariableToValue) == 0);
 	}
 
 	public void testCreateUserVariable() {
